scoring-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pyodbc==5.0.1
pandas==2.1.3
python-dotenv==1.0.0
openpyxl==3.1.2
